Fix: Use `unordered_multimap::iterator` instead of `hash_map::iterator`.
In ApiReader.cpp, ApiReader::clearAll() is iterating through apiList, which is multi.

In ApiReader.h, ApiReader::getScoredApi() should be declared as taking multi to match its definition in ApiReader.cpp,
as it's taking iterators into apiList.

This was compiling despite the single/multi mismatch because in MSVC's implementation, for both legacy hash_map and
Standard unordered_map, the containers have the same iterator types regardless of single/multi, but it's still a conformance
issue and should be fixed.